updateGeminiCLI/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
Flask-JWT-Extended
Flask-Mail
Flask-Cors
python-dotenv
